+++ semantic_router/__init__.py
@@ -1,3 +1,4 @@
-from .layer import RouteLayer, HybridRouteLayer
+from .hybrid_layer import HybridRouteLayer
+from .layer import RouteLayer
 
 __all__ = ["RouteLayer", "HybridRouteLayer"]
